+++ docs/content/changelog.md
@@ -15,6 +15,7 @@ title = "Changelog"
 - add `c4ddev.GetClipMapHandle()`
 - add `c4ddev.BlitClipMap()`
 - add `c4ddev.am` module
+- fix `c4ddev pypkg` create directory if it doesnt exist
 
 ## v0.1.5